Move getOfferingProductTags to service layer
