+++ swatch-contracts/src/main/java/com/redhat/swatch/contract/service/ContractService.java
@@ -60,15 +60,18 @@ public class ContractService {
   private final ContractRepository contractRepository;
   private final ContractMapper mapper;
   private final SubscriptionSyncResource syncResource;
+  private final SubscriptionSyncService syncService;
   @Inject @RestClient PartnerApi partnerApi;
 
   ContractService(
       ContractRepository contractRepository,
       ContractMapper mapper,
-      SubscriptionSyncResource syncResource) {
+      SubscriptionSyncResource syncResource,
+      SubscriptionSyncService syncService) {
     this.contractRepository = contractRepository;
     this.mapper = mapper;
     this.syncResource = syncResource;
+    this.syncService = syncService;
   }
 
   /**
@@ -456,7 +459,7 @@ public class ContractService {
           var sku = rhEntitlements.get(0).getSku();
           prevEntity.setSku(sku);
           prevEntity.setSubscriptionNumber(subscription);
-          OfferingProductTags productTags = syncResource.getOfferingProductTags(sku);
+          OfferingProductTags productTags = syncService.getOfferingProductTags(sku);
           if (Objects.nonNull(productTags.getData())
               && Objects.nonNull(productTags.getData().get(0))) {
             prevEntity.setProductId(productTags.getData().get(0));
@@ -521,7 +524,7 @@ public class ContractService {
       entity.setSku(sku);
       log.trace("Call swatch api to get producttags by sku {}", sku);
       try {
-        OfferingProductTags productTags = syncResource.getOfferingProductTags(sku);
+        OfferingProductTags productTags = syncService.getOfferingProductTags(sku);
         if (Objects.nonNull(productTags)
             && Objects.nonNull(productTags.getData())
             && !productTags.getData().isEmpty()
